Changed integer to bigInteger in migration

+++ server/migrations/20170520110722_create-expenses.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('hw_values', function (table) {
     table.increments()
     table.string('name')
-    table.integer('value')
+    table.bigInteger('value')
     table.date('date')
   })
 };
